+++ tokamaksrc/src/constraint.h
@@ -59,7 +59,7 @@ typedef neCollection<_neConstraint>::itemType neConstraintHandle;
 typedef struct neLimitState neLimitState;
 
 struct neLimitState {
-    s32 limitType;
+    int32_t limitType;
 
     neBool applyLimitImpulse;
 
@@ -83,7 +83,7 @@ struct neLimitState {
 
     _neConstraint *constr;
 
-    void Reset(_neConstraint *c, s32 ltype = 0) {
+    void Reset(_neConstraint *c, int32_t ltype = 0) {
         constr = c;
 
         applyLimitImpulse = false;
@@ -156,7 +156,7 @@ public:
     neConstraintHandle bodyAHandle;
     neConstraintHandle bodyBHandle;
 
-    s32 pointCount;
+    size_t pointCount;
 
     neJointPoint cpointsA[2];
     neJointPoint cpointsB[2];
@@ -172,7 +172,7 @@ public:
 
     f32 accuracy;
 
-    s32 iteration;
+    int32_t iteration;
 
     f32 jointLength;
 
@@ -214,7 +214,7 @@ public:
 
     void AddToRigidBody();
 
-    neController *AddController(neJointControllerCallback *jc, s32 period);
+    neController *AddController(neJointControllerCallback *jc, int32_t period);
 
     void BeginIterateController();
 
@@ -264,13 +264,13 @@ public:
 
     neBool solved;
 
-    s32 flag;
+    int32_t flag;
 
     neCollection<neRigidBodyBase> bodies;
 
     neConstraintHeader() { Reset(); }
 
-    void AddToSolver(f32 &epsilon, s32 &iteration);
+    void AddToSolver(f32 &epsilon, int32_t &iteration);
 
     void Reset() {
         head = NULL;
